requirements.txt
cupy-cuda12x
numpy
pandas
matplotlib
scipy
scikit-learn 